Updated Colony
The scrollbar hotkey finally works as fully intended without any noticeable issues. 

Earlier I released the scrollbar update but it had an issue, a quick hotfix was released but retracted due to the discovery of more problems. Issues before the fix included:
- When your unit group died, you could continue to select them and other of the same units from beyond the grave
- When your base died, the match would never end, as if it were still alive
- You could group select units that weren't of your own (this is a bug in v6.1 too, and not good for performance. It should now only apply to your own units!)
- Ghosts appeared in your living room

Now that it's fixed, you can enjoy it to its full glory! This update is not in the Colony Player yet, this will come sometime within 24-48 hours of this commit. Have fun to anyone that reads this... 

👻